Added investment in summary response

+++ src/main/java/com/alok/spring/response/GetMonthlySummaryResponse.java
@@ -21,6 +21,6 @@ public class GetMonthlySummaryResponse {
         private Double expenseAmount;
         private Double transferAmount;
         private Double savingAmount;
-        private Double investmentAmount;
+        private Long investmentAmount;
     }
 }
